renamed indexing submodule
