+++ python/SMQTK/indexing/naive_bayes.py
