+++ python/SMQTK/indexing/svm_hik.py
@@ -11,12 +11,12 @@ import logging
 import multiprocessing.pool
 import numpy
 import os.path as osp
-import svm
+import svm_hik
 import svmutil
 
 from EventContentDescriptor.iqr_modules import svmtools
 
-from SMQTK.Indexers import Indexer
+from SMQTK.indexing import Indexer
 from SMQTK.utils import SimpleTimer
 from SMQTK.utils.distance_functions import histogram_intersection_distance
 
@@ -368,7 +368,7 @@ class SVMIndexer_HIK (Indexer):
             self._feature_mat[list(self._uid2idx_map[uid] for uid in uid_list), :]
 
         self.log.debug("Creating SVM problem")
-        svm_problem = svm.svm_problem(labels, train_features.tolist())
+        svm_problem = svm_hik.svm_problem(labels, train_features.tolist())
         self.log.debug("Creating SVM model")
         w1_weight = max(1.0, len(neg_ids)/float(len(pos_ids)))
         svm_model = svmutil.svm_train(svm_problem,
